login auth working

+++ routes/days.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const passport = require('passport');
 
 const router = express.Router();
 
@@ -8,14 +9,19 @@ const Trip = require('../models/trip');
 
 const mongoose = require('mongoose');
 
+
+router.use('/', passport.authenticate('jwt', { session: false, failWithError: true }));
+
 router.get('/', (req, res, next) => {
-  Day.find()
+  const userId = req.user.id;
+  Day.find({userId})
     .then(result => result ? res.json(result) : next())
     .catch(err => next(err));
 });
 
 router.get('/:id' , (req, res, next) => {
   const {id} = req.params;
+  const userId = req.user.id;
   
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -23,7 +29,7 @@ router.get('/:id' , (req, res, next) => {
     return next(err);
   }
   
-  Day.findOne({_id: id})
+  Day.findOne({_id: id, userId})
     .populate('plans')
     .then(result => {
       if (result) {
@@ -39,8 +45,9 @@ router.get('/:id' , (req, res, next) => {
 
 router.post('/', (req, res, next) => {
   const { content, plans = [] } = req.body;
+  const userId = req.user.id;
 
-  const newDay = { content, plans };
+  const newDay = { content, plans, userId };
   
   Day.create(newDay)
     .then(result => {
@@ -57,6 +64,7 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { content, plans = [] } = req.body;
   const {id} = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -66,7 +74,7 @@ router.put('/:id', (req, res, next) => {
 
   const updateDay = {content, plans};
 
-  Day.findOneAndUpdate({_id:id}, updateDay, { new: true })
+  Day.findOneAndUpdate({_id:id, userId}, updateDay, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -80,6 +88,7 @@ router.put('/:id', (req, res, next) => {
 
 router.delete('/:id', (req, res, next) => {
   const {id} = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -92,7 +101,7 @@ router.delete('/:id', (req, res, next) => {
     { $pull: { days: id } }
   );
 
-  const dayDeletePromise = Day.deleteOne({_id:id});
+  const dayDeletePromise = Day.deleteOne({_id:id, userId});
 
   Day.findOne({_id:id})
     .then(day => day.plans)
